Double precision for maths in 2d Geometries

+++ Src/Geometry/2d/Geom2d_abstr.cpp
@@ -19,31 +19,31 @@
 #include "Point_abstr.hpp"
 
 template<class Geom>
-glm::vec2 Geom2d_abstr<Geom>::at(float t)
+glm::dvec2 Geom2d_abstr<Geom>::at(double t)
 {
 	return static_cast<Geom*>(this)->at_geom(t);
 }
 
 template<class Geom>
-bool Geom2d_abstr<Geom>::within(glm::vec2 top_left, glm::vec2 bottom_right, bool contained)
+bool Geom2d_abstr<Geom>::within(glm::dvec2 top_left, glm::dvec2 bottom_right, bool contained)
 {
 	return static_cast<Geom*>(this)->within_geom(top_left, bottom_right, contained);
 }
 
 template<class Geom>
-float Geom2d_abstr<Geom>::closest_to_point_interp(glm::vec2 const& pt)
+double Geom2d_abstr<Geom>::closest_to_point_interp(glm::dvec2 const& pt)
 {
 	return static_cast<Geom*>(this)->closest_to_point_interp_geom(pt);
 }
 
 template<class Geom>
-glm::vec2 Geom2d_abstr<Geom>::closest_to_point(glm::vec2 const& pt)
+glm::dvec2 Geom2d_abstr<Geom>::closest_to_point(glm::dvec2 const& pt)
 {
 	return static_cast<Geom*>(this)->closest_to_point_geom(pt);
 }
 
 template<class Geom>
-float Geom2d_abstr<Geom>::dist_to_point(glm::vec2 const& pt)
+double Geom2d_abstr<Geom>::dist_to_point(glm::dvec2 const& pt)
 {
     return static_cast<Geom*>(this)->dist_to_point_geom(pt);
 }
