+++ Src/Geometry/2d/Line_abstr.cpp
@@ -26,34 +26,34 @@ Line_abstr<Impl>::Line_abstr()
 }
 
 template<typename Impl>
-glm::vec2 Line_abstr<Impl>::posA()
+glm::dvec2 Line_abstr<Impl>::posA()
 {
 	return static_cast<Impl*>(this)->posA_impl();
 }
 template<typename Impl>
-glm::vec2 Line_abstr<Impl>::posB()
+glm::dvec2 Line_abstr<Impl>::posB()
 {
 	return static_cast<Impl*>(this)->posB_impl();
 }
 template<typename Impl>
-float Line_abstr<Impl>::length()
+double Line_abstr<Impl>::length()
 {
 	return glm::distance(posA(), posB());
 }
 template<typename Impl>
-float Line_abstr<Impl>::length2()
+double Line_abstr<Impl>::length2()
 {
 	return glm::distance2(posA(), posB());
 }
 
 template<typename Impl>
-glm::vec2 Line_abstr<Impl>::at_geom(float t)
+glm::dvec2 Line_abstr<Impl>::at_geom(double t)
 {
 	return (posA() - t * as_vec());
 }
 
 template<typename Impl>
-bool Line_abstr<Impl>::within_geom(glm::vec2 top_left, glm::vec2 bottom_right, bool contained)
+bool Line_abstr<Impl>::within_geom(glm::dvec2 top_left, glm::dvec2 bottom_right, bool contained)
 {
 	// If the whole line has to be contained, return true if both points are within the selection
 	if(contained) {
@@ -62,19 +62,19 @@ bool Line_abstr<Impl>::within_geom(glm::vec2 top_left, glm::vec2 bottom_right, b
 
 	// If only a part has to be contained, return if either point is contained or 
 	// if the line crosses  any of the four lines of the selection rectangle
-	glm::vec2 top_right(bottom_right.x, top_left.y), bottom_left(top_left.x, bottom_right.y);
+	glm::dvec2 top_right(bottom_right.x, top_left.y), bottom_left(top_left.x, bottom_right.y);
 	return point_within(posA(), top_left, bottom_right) || point_within(posB(), top_left, bottom_right)
 			|| intersects(top_left, top_right) || intersects(bottom_left, bottom_right)
 			|| intersects(top_left, bottom_left) || intersects(top_right, bottom_right);
 }
 template<typename Impl>
-float Line_abstr<Impl>::closest_to_point_interp_geom(glm::vec2 const& pt)
+double Line_abstr<Impl>::closest_to_point_interp_geom(glm::dvec2 const& pt)
 {
-	float len2 = length2();
+	double len2 = length2();
 	if(len2 == 0) {
 		return 0.0f;
 	}
-	float t = glm::dot(as_vec(), posA()-pt) / len2;
+	double t = glm::dot(as_vec(), posA()-pt) / len2;
 	if(t <= 0) { // First point
 		return 0.0f;
 	} else if(t >= 1) { // Second point 
@@ -84,7 +84,7 @@ float Line_abstr<Impl>::closest_to_point_interp_geom(glm::vec2 const& pt)
 	}
 }
 template<typename Impl>
-float Line_abstr<Impl>::dist_to_point_geom(glm::vec2 const& pt)
+double Line_abstr<Impl>::dist_to_point_geom(glm::dvec2 const& pt)
 {
 	return glm::distance(closest_to_point_geom(pt), pt);
 }
@@ -96,7 +96,7 @@ bool Line_abstr<Impl>::intersects(Line_abstr* l)
 	return ccw(posA(), l->posA(), l->posB()) != ccw(posB(), l->posA(), l->posB()) && ccw(posA(), posB(), l->posA()) != ccw(posA(), posB(), l->posB());
 }
 template<typename Impl>
-bool Line_abstr<Impl>::intersects(glm::vec2 a, glm::vec2 b)
+bool Line_abstr<Impl>::intersects(glm::dvec2 a, glm::dvec2 b)
 {
 	// Better explanation than I could give
 	// https://bryceboe.com/2006/10/23/line-segment-intersection-algorithm/
@@ -104,18 +104,18 @@ bool Line_abstr<Impl>::intersects(glm::vec2 a, glm::vec2 b)
 }
 
 template<typename Impl>
-glm::vec2 Line_abstr<Impl>::as_vec()
+glm::dvec2 Line_abstr<Impl>::as_vec()
 {
 	return posA() - posB();
 }
 template<typename Impl>
-bool Line_abstr<Impl>::ccw(glm::vec2 A, glm::vec2 B, glm::vec2 C)
+bool Line_abstr<Impl>::ccw(glm::dvec2 A, glm::dvec2 B, glm::dvec2 C)
 {
 	return (C.y-A.y) * (B.x-A.x) > (B.y-A.y) * (C.x-A.x);
 }
 
 
-Line::Line(glm::vec2 a, glm::vec2 b)
+Line::Line(glm::dvec2 a, glm::dvec2 b)
 	: mA(new Point(a))
 	, mB(new Point(b))
 {
